// src/sdk/generated/nfl.d.ts
export interface paths {
// ...
    "/v3/nfl/scores/{format}/AllTeams": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Team Profiles - All
         * @description Full team information: team name and city, conference and division, colors, coaching and scheme info. Also contains basic fantasy info such as IDs as well as full stadium data. This endpoint returns all teams regardless of current active status.
         */
        get: operations["nfl_v3_scores_team_profiles___all"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}