// src/sdk/generated/cbb.d.ts
export interface paths {
// ...
    "/v3/cbb/scores/{format}/teams": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Team Profiles - All
         * @description Full team information: team name and city, school, conference, and colors. Also contains full stadium data. This endpoint returns all teams regardless of current active status.
         */
        get: operations["cbb_v3_scores_team_profiles___all"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}