// src/sdk/generated/cwbb.d.ts
export interface paths {
// ...
    "/v3/cwbb/scores/{format}/Teams": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Team Profiles - All
         * @description Full team information: school, team name and city, conference and division, and colors.
         */
        get: operations["cwbb_v3_scores_team_profiles___all"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}